+++ TarSCM/scm/__init__.py
@@ -1,0 +1,5 @@
+from git import git
+from bzr import bzr
+from hg import hg
+from svn import svn
+from tar import tar
