+++ tests/test.py
@@ -9,6 +9,8 @@ import shutil
 import sys
 import unittest
 
+
+
 from gittests import GitTests
 from svntests import SvnTests
 from hgtests  import HgTests
@@ -16,6 +18,7 @@ from bzrtests import BzrTests
 from testenv import TestEnvironment
 from unittestcases import UnitTestCases
 from snapcraft import SnapcraftTestCases
+
 
 if __name__ == '__main__':
     suite = unittest.TestSuite()
@@ -42,6 +45,8 @@ if __name__ == '__main__':
         #   suite.addTest(HgTests('test_version_versionformat'))
         #   suite.addTest(HgTests('test_versionformat_dateYYYYMMDD'))
         test_class = GitTests
+        #test_class = SnapcraftTestCases
+        #test_class = UnitTestCases
         to_run = {}
         for arg in sys.argv[1:]:
             m = re.match('^/(.+)/$', arg)
